+++ src/components/ui/NavButton.tsx
@@ -13,7 +13,7 @@ export function NavButton({ active = false, expanded, children, onClick, as = 'b
   const Tag = as
   return (
     <Tag
-      className={`nav-btn rounded-md py-2.5 lg:py-2 2xl:py-3 3xl:py-4 4xl:py-5 px-2 3xl:px-4 4xl:px-5 ${active ? 'active' : ''}`}
+      className={`nav-btn ${active ? 'active' : ''}`}
       aria-current={active ? 'page' : undefined}
       aria-expanded={expanded}
       onClick={onClick}